% Generated by build script
\name{acorps_cev_national}
\alias{acorps_cev_national}
\title{CEV national rates of civic engagement measures (2017-2023)}
\usage{
acorps_cev_national()
}
\arguments{
\item{limit}{Max rows (default 100)}
\item{token}{Optional Socrata app token}
}
\value{tibble: year, national_formal_volunteering,}
\description{CEV national rates of civic engagement measures (2017-2023)}
