% Generated by build script
\name{acorps_columns}
\alias{acorps_columns}
\title{Get column metadata for an AmeriCorps dataset}
\usage{
acorps_columns()
}
\arguments{
\item{view_id}{Socrata 4x4 view identifier (e.g. "yie5-ur4v")}
}
\value{tibble: field_name, name, datatype, description}
\description{Get column metadata for an AmeriCorps dataset}
