% Generated by build script
\name{acorps_list}
\alias{acorps_list}
\title{List all AmeriCorps open datasets}
\usage{
acorps_list()
}
\arguments{
\item{category}{Optional filter: "deployments", "education_award",}
}
\value{tibble: view_id, name, category}
\description{List all AmeriCorps open datasets}
